# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bank API",
    "description": "API Core para la gestión de usuarios, cuentas y movimientos",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Pepe"
          },
          "surname": {
            "type": "string",
            "example": "Gonzalez"
          },
          "role": {
            "type": "string",
            "example": "USER",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "password": {
            "type": "string",
            "example": "a123456789"
          },
          "email": {
            "type": "string",
            "example": "pepe@gmail.com"
          },
          "age": {
            "type": "number",
            "example": 33
          },
          "city": {
            "type": "string",
            "example": "SEG"
          }
        },
        "required": [
          "name",
          "surname",
          "role",
          "password",
          "email",
          "age",
          "city"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Pepe"
          },
          "surname": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "example": "USER",
            "enum": [
              "USER",
              "ADMIN"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "age": {
            "type": "number"
          },
          "city": {
            "type": "string"
          },
          "fullname": {
            "type": "string",
            "example": "Pepe Gonzalez"
          }
        },
        "required": [
          "id",
          "name",
          "surname",
          "role",
          "createdAt",
          "email",
          "age",
          "city",
          "fullname"
        ]
      },
      "BadRequest": {
        "type": "object",
        "properties": {
          "url": {
            "type": "string",
            "example": "/*"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string",
            "example": "2021-04-30T15:12:23.562Z"
          },
          "message": {
            "example": [
              "* must be uppercase",
              "* must be a valid enum value"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "statusCode": {
            "type": "number",
            "example": 400
          }
        },
        "required": [
          "url",
          "timestamp",
          "message",
          "statusCode"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  },
  "paths": {
...
    "/user": {
      "get": {
        "operationId": "UsersController_getAll",
        "parameters": [
          {
            "name": "ids",
            "required": false,
            "in": "query",
            "example": [
              1,
              2
            ],
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "example": "pepe@gmail.com",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UsersController_getOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_delete",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "user"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  }
}